+++ core/CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add new protocol version 1.18.
 - Add `SubscribeService`, `SubscribeServiceReply` and `UnsubscribeService` messages.
 - Add `subscribe_all` field to `ServiceInfo`.
-- Add `SubscribeAllEvents`, `SubscribeAllEventsReply` and `UnsubscribeAllEvents` messages.
+- Add `SubscribeAllEvents`, `SubscribeAllEventsReply`, `UnsubscribeAllEvents` and
+  `UnsubscribeAllEventsReply` messages.
 
 ### Changed